Updated Prefectures Redux
Updated prefectures selector and reducer to pass the selected prefecture as an object into state. The prefecture in state can be consumed by the spinner to display the name of the prefecture that it's fetching cities for. Noone will ever see this because the spinner appears/disappears so quick but I'm glad it's there.